# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales CRM API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User management
  - name: customers
    description: Customer operations
  - name: leads
    description: Lead/interaction operations
  - name: dashboard
    description: Dashboard and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID with full lead history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer with lead history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads (scoped to current user unless admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
        - name: followUpToday
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadWithDetails"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead/interaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent lead activity
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadWithDetails"

  /dashboard/status-breakdown:
    get:
      operationId: getStatusBreakdown
      tags: [dashboard]
      summary: Get leads grouped by status
      responses:
        "200":
          description: Status breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"

  /leads/{id}/send-followup:
    post:
      operationId: sendFollowupEmail
      tags: [leads]
      summary: Trigger a follow-up email for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        staffId:
          type: integer
        role:
          type: string
          enum: [admin, sales]
        createdAt:
          type: string
      required:
        - id
        - email
        - staffId
        - role

    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        staffId:
          type: integer
        role:
          type: string
          enum: [admin, sales]
      required:
        - email
        - password
        - staffId
        - role

    UpdateUserBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        staffId:
          type: integer
        role:
          type: string
          enum: [admin, sales]

    Customer:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: ["string", "null"]
        streetAddress:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - companyName
        - contactName

    CustomerProfile:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: ["string", "null"]
        streetAddress:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        createdAt:
          type: string
        leads:
          type: array
          items:
            $ref: "#/components/schemas/LeadWithDetails"
      required:
        - id
        - companyName
        - contactName
        - leads

    CreateCustomerBody:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
      required:
        - companyName
        - contactName

    UpdateCustomerBody:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string

    Lead:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        userId:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [New, Contacted, Qualified, Proposal, Won, Lost]
        followUpDate:
          type: ["string", "null"]
        dateKey:
          type: ["integer", "null"]
        metadata:
          type: ["object", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerId
        - userId
        - status
        - createdAt

    LeadWithDetails:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        userId:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [New, Contacted, Qualified, Proposal, Won, Lost]
        followUpDate:
          type: ["string", "null"]
        dateKey:
          type: ["integer", "null"]
        metadata:
          type: ["object", "null"]
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - customerId
        - userId
        - status
        - createdAt
        - customer
        - user

    CreateLeadBody:
      type: object
      properties:
        customerId:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [New, Contacted, Qualified, Proposal, Won, Lost]
        followUpDate:
          type: string
        metadata:
          type: object
      required:
        - customerId

    UpdateLeadBody:
      type: object
      properties:
        notes:
          type: string
        status:
          type: string
          enum: [New, Contacted, Qualified, Proposal, Won, Lost]
        followUpDate:
          type: ["string", "null"]
        metadata:
          type: object

    DashboardSummary:
      type: object
      properties:
        totalLeads:
          type: integer
        myLeads:
          type: integer
        followUpsToday:
          type: integer
        totalCustomers:
          type: integer
        wonLeads:
          type: integer
        newLeads:
          type: integer
      required:
        - totalLeads
        - myLeads
        - followUpsToday
        - totalCustomers
        - wonLeads
        - newLeads

    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count