# lib/api-zod/src/generated/types/createLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLeadBodyMetadata } from "./createLeadBodyMetadata";
import type { CreateLeadBodyStatus } from "./createLeadBodyStatus";

export interface CreateLeadBody {
  customerId: string;
  notes?: string;
  status?: CreateLeadBodyStatus;
  followUpDate?: string;
  metadata?: CreateLeadBodyMetadata;
}