# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalLeads: number;
  myLeads: number;
  followUpsToday: number;
  totalCustomers: number;
  wonLeads: number;
  newLeads: number;
}