# lib/api-zod/src/generated/types/leadWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";
import type { LeadWithDetailsMetadata } from "./leadWithDetailsMetadata";
import type { LeadWithDetailsStatus } from "./leadWithDetailsStatus";
import type { User } from "./user";

export interface LeadWithDetails {
  id: string;
  customerId: string;
  userId: string;
  /** @nullable */
  notes?: string | null;
  status: LeadWithDetailsStatus;
  /** @nullable */
  followUpDate?: string | null;
  /** @nullable */
  dateKey?: number | null;
  /** @nullable */
  metadata?: LeadWithDetailsMetadata;
  createdAt: string;
  customer: Customer;
  user: User;
}